Enable search results with dates to be "snippeted" for search results
Rollbar: https://rollbar.com/dxw/tna-caselaw-public-ui/items/26/

This very obscure error was occurring when users searched for judgments with
to and from dates as part of the search parameters. The `snippet` functionality
in Marklogic was not able to handle search results from date searches due to
missing the `akn` namespace when formatting results.

Fix supplied by jurisdatum

Co-Authored-By: jurisdatum <jim@jurisdatum.com>