Amend image extension rexeg to include line end
Suggested by @dragon-dxw:  the previous regex could conceivably match (for
example) `.svg.exe`. Amend the regex to ensure we're looking for the image
file extension at the end of the string.

(This commit is an almost-clone of 8eec793df4595fc8acae8c7b379fd611e21e1f46
which was reverted; the regex was not valid for XSLT)