OpenAPI documentation passes validation
Optional path strings are not a part of the OpenAPI spec but
have been a long-running request.

https://github.com/OAI/OpenAPI-Specification/issues/93

So we tell a small fib in the spec and say these paths are required,
(because they *have* to be required in the spec) but then
have a comment pointing out that /atom.xml etc are valid paths.

Additionally we add some scripts to install an OpenAPI validator
and to validate the two YAML documents. There is documentation
for doing this with pre-commit but I couldn't get that to work.