Add transformation sort order for search
This is so we can sort by manifestation order for the atom feed
We have to look for the transformation date specifically because
there can be multiple dates in the Manifestation section of the document
for FRBRManifestation.

If we want to sort by Manifestation date for downstream consumers,
we need to index it.

We thought that invalid-values: ignore is required on that index because there is no manifestation date for newly ingested documents, causing errors like https://rollbar.com/dxw/tna-caselaw-ingester/items/42/?utm_campaign=new_item_message&utm_medium=slack&utm_source=rollbar-notification , but this also masked an issue where we used a date instead of a dateTime which
lead to it indexing everything as "not a date" (we think)