Re-enable rendering of non-websafe images
After discussion with the editorial team, we have decided it is more helpful
for non-websafe images to show as broken images in the HTML rengering of a
judgment, than for nothing to appear. If we show a broken image, then the
editorial team can know to send the judgment back to the clerk, and ask for
images to be supplied in an approved format.

https://trello.com/c/TLYCwwjX