Add a new privilege and role allowing users to view unpublished documents
Trello: https://trello.com/c/o3Wbto2N/263-implement-preview-permissions

There is a requirement with the privileged API to grant some users the ability
to view unpublished judgments, depending on their user account.

The thought behind this is to add a new `caselaw-unpublished-reader` role
which is similar to the existing `caselaw-reader` role, but indicates the users
with this role can view unpublished documents.

Further down the line, we can add an endpoint to `ds-caselaw-custom-api-client`
where we can pass in a useer name, and check to see if they have this role/
privilege.