Create failover replicas for the caselaw-content databases
By default a marklogic cluster does not deal with losing forests from a
database very well. If it cant find a forest due to a machine being missing it
will stop responding to queries.

This sets and creates failover replicas of the forests needed for the
application so that losing a forest on a machine should not impact the
clusters health too much.

https://github.com/marklogic-community/ml-gradle/wiki/Creating-forests#replica-forest-creation