Rename judgment0.xsl and judgment2.xsl to have more descriptive names
Rename judgment0 to `as-handed-down.xsl` to indicate it is the XSL Transform
which will render the judgment XML as close to "as handed down" by the judge
as possible.

Rename judgment2 to `accessible-html.xsl' to indicate it is the XSL Transform
while renders a accessible HTML version of the judgment.

When uploading these to Marklogic via Gradle, the existing judgment0.xsl and
judgment2.xsl in caselaw-modules were not overwritten, so it is safe to deploy
this before we have made the appropriate edits to the Custom Marklogic API
Client.