Version search xquery (part of Allow multicourt in search query)
The XQuery now expects a list of strings rather than a single string
for the 'court' parameter, which is a breaking change. As such,
we've decided to instead create a new separate version of search.xqy
so that the Public/Editor UIs do not need to move in lockstep with the
custom-api-client and public-access-service repos, and can continue
to use the older, not-explicitly-versioned search.xqy

We should probably delete search.xqy when we are done.

This will become version 2 of the public-access-service repo.