Add dls-internal, dls-user role to all caselaw- roles
All caselaw-* roles inherit from caselaw-reader.

All caselaw-* roles need the dls-internal role, in order to be able to search
within the document properties.

We were having an issue where the caselaw roles were unable to see any
judgments, without being assigned the admin role. We do not want all users
to have the admin role, for security reasons.

It transpires that because the judgments are ingested with the dls-internal
role, any user roles which do not inherit from this role are unable to
search for them.