Add the `xdbc-eval` and `xdbc-invoke` privileges to caselaw-reader
These two privileges are needed to cal lthe `eval` and `invoke` endpoints
on the Marklogic REST API.

These privileges had been set via the Marklogic admin console on staging, but
not copied over to the Gradle config.

Todo: Find a better way to make sure any changes made in the consoles are
recorded in the Gradle config!