Get metadata when searching via TDR
https://trello.com/c/k4qWwcK3/140

I do not fully understand what is happening here, but I think I
understand it better than when I did before.

The crux of the bug is, I think, twofold:

When searching with a TDR- query, it searched for the search query
in the selected XML (via `$query1`) and explicitly in the
`transfer-consignment-reference` property of the relevant document.

The selected XML (`$scope`, the `<fragment-scope/>`) was the
**properties** of the document, not the judgment XML itself, hence
the title was missing (because that is part of the judgment XML,
not a metadata property).

But: if we change the selected XML to be the document, the search
fails, because the search query (`TDR-2022-XML` for example) does
not appear within the judgment XML.

So we have to fix it in two places.

This all makes perfect sense... if it wasn't for why it works fine
and has titles if `$order` is `updated` and `$scope` is set to
`properties`...