Add repository dashboard
Since this project has so many repositories, create a central list and
use shields.io to extract the status of various aspects to give a
high-level overview. At the moment, this is number of open PRs and
latest release semver.

There are enough repositories, and the combined Markdown table/image
formats are messy enough, that it's easiest to generate this using a
quick and dirty Python script.