Tidy README and add explicit CONTRIBUTING documentation
This cleans up the README, fixes a couple of typographical glitches, and
links to the wiki for non-technical documentation. It also moves
contributing data into its own file, so GitHub can pick it up and flag
it to people engaging with the project for the first time.

This _explicitly_ breaks the Readme Standard, but since that hasn't been
maintained for a few years and seems quite heavy-handed I personally
prefer things being more usable than blindly following a spec.