Run pre-commit
This tidies up the formatting of our Markdown and yml files.

One error was exposed: a duplicate version key in the public API spec.

This _does_ reformat the autogenerated parts of documentation spat out
by `build_repo_lists`, this is intentional so that any updates as a
result of this also have to be prettified before they can be committed.